Redesign sorting and 'show more'
Redesigned the way that show more and sorting are handled. Instead of remaking the HTML elements, which would delete the listeners attached and therefore reset the favorites set by the user, the attempt now is to sort the same existing HTML elements, and hide all the elements that aren't called for until 'show more' is clicked the required amount of times.